+++ src/main/java/tornadofx/ItemControls.kt
@@ -1,6 +1,7 @@
 package tornadofx
 
 import com.sun.javafx.scene.control.skin.TableRowSkin
+import javafx.application.Platform
 import javafx.beans.binding.Bindings
 import javafx.beans.property.*
 import javafx.beans.value.ChangeListener
@@ -299,7 +300,14 @@ fun <S> TableColumn<S, Boolean?>.useCheckbox(editable: Boolean = true): TableCol
     return this
 }
 
-class CheckBoxCell<S>(val editable: Boolean) : TableCell<S, Boolean?>() {
+class CheckBoxCell<S>(val makeEditable: Boolean) : TableCell<S, Boolean?>() {
+    init {
+        if (makeEditable) {
+            isEditable = true
+            tableView?.isEditable = true
+        }
+    }
+
     override fun updateItem(item: Boolean?, empty: Boolean) {
         super.updateItem(item, empty)
         style { alignment = Pos.CENTER }
@@ -309,16 +317,20 @@ class CheckBoxCell<S>(val editable: Boolean) : TableCell<S, Boolean?>() {
         } else {
             graphic = CheckBox().apply {
                 val model = tableView.items[index]
-                val prop = tableColumn.cellValueFactory.call(TableColumn.CellDataFeatures(tableView, tableColumn, model)) as ObjectProperty
-                isEditable = editable
+                val prop = tableColumn.cellValueFactory.call(TableColumn.CellDataFeatures(tableView, tableColumn, model)) as Property
 
-                if (editable) {
+                if (makeEditable) {
                     selectedProperty().bindBidirectional(prop)
+                    selectedProperty().onChange {
+                        tableView.edit(tableRow.index, tableColumn)
+                        commitEdit(isSelected)
+                    }
                 } else {
                     disableProperty().set(true)
                     selectedProperty().bind(prop)
                 }
             }
+
         }
     }
 }
@@ -382,6 +394,31 @@ inline fun <reified S, T> TableView<S>.column(title: String, prop: KProperty1<S,
     addColumnInternal(column)
     op?.invoke(column)
     return column
+}
+
+/**
+ * Add a global edit commit handler to the TableView. You avoid assuming the responsibility
+ * for writing back the data into your domain object and can consentrate on the actual
+ * response you want to happen when a column commits and edit.
+ */
+fun <S> TableView<S>.onEditCommit(onCommit: TableColumn.CellEditEvent<S, Any>.(S) -> Unit) {
+    fun addEventHandlerForColumn(column: TableColumn<S, *>) {
+        column.addEventHandler(TableColumn.editCommitEvent<S, Any>()) { event ->
+            // Make sure the domain object gets the new value before we notify our handler
+            Platform.runLater {
+                onCommit(event, event.rowValue)
+            }
+        }
+    }
+
+    columns.forEach(::addEventHandlerForColumn)
+
+    columns.addListener({ change: ListChangeListener.Change<out TableColumn<S, *>> ->
+        while (change.next()) {
+            if (change.wasAdded())
+                change.addedSubList.forEach(::addEventHandlerForColumn)
+        }
+    })
 }
 
 /**