+++ src/test/kotlin/tornadofx/tests/BindingTests.kt
@@ -8,6 +8,7 @@ import org.junit.Assert
 import org.junit.Test
 import org.testfx.api.FxToolkit
 import tornadofx.bind
+import tornadofx.select
 import java.text.DecimalFormat
 import java.text.DecimalFormatSymbols
 import java.util.*
@@ -33,6 +34,21 @@ class BindingTests {
         Assert.assertEquals("Hello World", uielement.text)
         elements.setAll("Hello", "Changed")
         Assert.assertEquals("Hello Changed", uielement.text)
+    }
+
+    @Test
+    fun nestedBinding() {
+        val father = Person("Mr Father", 50)
+        val stepFather = Person("Mr Step Father", 40)
+        val child = Person("Firstborn Child", 18)
+        child.parent = father
+
+        val fatherName = child.parentProperty().select(Person::nameProperty)
+        Assert.assertEquals("Mr Father", fatherName.value)
+        fatherName.value = "Mister Father"
+        Assert.assertEquals("Mister Father", father.name)
+        child.parent = stepFather
+        Assert.assertEquals("Mr Step Father", fatherName.value)
     }
 
 }