+++ CHANGELOG.md
@@ -5,12 +5,22 @@ All notable changes to this project will be documented in this file.
 ## [1.5.10-SNAPSHOT]
 
 - di delegate overload to support injecting a dependency by name (in addition to type)
+- `adhocFragment` and `adhocWindow` builders
+- `ObservableList<T>.onChange` to easy listening to change events from observable lists
+- `setInScope()` now uses correct KClass when entering the injectable into the components map
+- `ItemViewModel.isEmpty` boolean, complements `empty` property
+- `setStageIcon(icon)` will replace all existing icons with the supplied (https://github.com/edvin/tornadofx/issues/228)
+- `TableColumn.useCheckbox(editable = true)` now fires edit/commit events when value is changed
+- Create nested, observable, writable properties using the `observableValue.select()` function
+- ViewModel `bind` has optional parameter `forceObjectProperty` to avoid creating `IntegerProperty` for ints etc, so you can have nullable values
+- `TableView.onEditCommit()` handler fires when a cell is edited. No need to manage domain object value, just add your business logic
 - Fixed scope support. `DefaultScope(MyController::class)` or `MyController::class.scope(DefaultScope)`
 - TableColumn hasClass/addClass/removeClass/toggleClass supports type safe stylesheets
 - Lots of functions that earlier accepted Double now accept Number
 - TableView.enableCellEditing() makes table editable and enables cell selection
 - TableView.regainFocusAfterEdit() - make sure TableView doesn't look focus after cell edit
-- TableColumn.makeEditable(converter) - supply specific converter
+- TableColumn.makeEditable(converter) - supply specific converter for editable fields
+- TableColumn.converter(converter) - supply specific converter for read only text fields
 - TableColumn.makeEditable() supports BigDecimal
 - Added scope.set(injectable) as easier alternative to setInScope(injectable, scope)
 - tableview builder that takes `ObservableValue<ObservableList<T>>`, supporting automatic rebind when items change